+++ custom_components/ha_tracker/www/js/main.js
@@ -2,12 +2,12 @@
 // MAIN
 //
 
-import {isAdmin} from './globals.js';
-import {map, initMap, fitMapToAllDevices} from './map.js';
+import {isConnected, updateConnection, updateAdmin} from './globals.js';
+import {map, initMap} from './map.js';
 import {load, showWindowOverlay, hideWindowOverlay, configureConsoleLogging} from './utils.js';
-import {fetchAuthCallback,fetchConnection, fetchAdmin, fetchDevices, fetchZones, fetchPersons} from './fetch.js';
-import {updateDevicePerson, updateDeviceList, updateDeviceMarkers} from './devices.js';
-import {updateZonesTable, updateZoneMarkers} from './zones.js';
+import {authCallback} from './auth.js';
+import {updatePersons, fitMapToAllPersons} from './persons.js';
+import {updateZones} from './zones.js';
 import {initializeI18n, t} from './i18n.js';
 
 
@@ -18,20 +18,17 @@ document.addEventListener("DOMContentLoaded", async() => {
 
 		console.log("************** INICIANDO **************");
 	
-		await initializeI18n(); // Detecta el idioma y carga las traducciones
+		// Detecta el idioma y carga las traducciones
+		await initializeI18n(); 
 		
         // Manejar autenticación si hay un parámetro `code`
-		const urlParams = new URLSearchParams(window.location.search);
-        if (urlParams.has("code")) {
-            const code = urlParams.get("code");
-            await fetchAuthCallback(code);
-        }
+		await authCallback();
 
         // Inicializar la aplicación
         await init();
         await load();
     } catch (error) {
-        console.error("Error durante la autenticación o inicialización:", error);
+        console.error("Error durante la inicialización:", error);
     }
 });
 
@@ -43,7 +40,7 @@ async function init() {
         await update();
 
         // Zoom al conjunto de dispositivos
-        await fitMapToAllDevices();
+        await fitMapToAllPersons();
 
         // Inicia el ciclo de actualización
         startUpdateLoop();
@@ -69,23 +66,16 @@ function delay(ms) {
 
 async function update() {
     try {
-        const isConnected = await fetchConnection();
+		await updateConnection();
         if (!isConnected) {
             throw new Error("No está conectado");
-            return;
         }
 
         // Ejecutar funciones en orden y detenerse si ocurre un error
         try {
-            await fetchAdmin();
-            await fetchPersons();
-            await fetchDevices();
-            await updateDevicePerson();
-            await updateDeviceList();
-            await updateDeviceMarkers();
-            await fetchZones();
-            await updateZonesTable();
-            await updateZoneMarkers();
+            await updateAdmin();
+            await updatePersons();
+            await updateZones();
         } catch (error) {
             console.error("Error en una de las funciones:", error);
             throw new Error("Error en una de las funciones.");