+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+
+## [0.0.13] - 2025-01-30
+
+### Added
+- New screen with Home Assistant users
+- Adjust map zoom on startup
+- We changed the text of the filter and popups with the users
+- Added translations
+- Changes in css
+
+### Fixed
+- Check the connection with Home Assistant
+- The line **ha_tracker** of `configuration.yaml` in README.md
+
+
 ## [0.0.12] - 2025-01-27
 
 ### Added